% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_functions.R
\name{collapser}
\alias{collapser}
\title{Build Collapses}
\usage{
collapser(df, variable, collapses)
}
\arguments{
\item{df}{dataframe to edit}

\item{variable}{string variable/column name to edit}

\item{collapses}{list of collapses with "label"=c(inputs).  use "@auto" to auto-generate labels.}
}
\value{
collapsed dataframe
}
\description{
Combine values in a given column to create collapsed categories
}
