% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_functions.R
\name{pretty_col_name}
\alias{pretty_col_name}
\title{Convert machine-readable tab column names to human-readable names.}
\usage{
pretty_col_name(col_name)
}
\arguments{
\item{col_name}{string column name}
}
\value{
string updated column name
}
\description{
Convert machine-readable tab column names to human-readable names.
}
