% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_functions.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder dataframe columns to display in logical order for tabs.}
\usage{
reorder_columns(df)
}
\arguments{
\item{df}{dataframe to edit}
}
\value{
updated dataframe
}
\description{
Reorder dataframe columns to display in logical order for tabs.
}
