% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_functions.R
\name{tab}
\alias{tab}
\title{Tabs - Unquoted Input}
\usage{
tab(df, v1, v2 = NULL, weight_var = NULL, sdesign = NULL,
  nsize = FALSE, ci = FALSE, to_factor = TRUE)
}
\arguments{
\item{df}{input dataframe}

\item{v1}{name of 1st variable/column to tabulate}

\item{v2}{name of optional 2nd variable/column to tablulate.  DEFAULT NULL}

\item{weight_var}{name of weight variable/column.  DEFAULT NULL}

\item{sdesign}{survey package surveydesign object.  DEFAULT NULL}

\item{nsize}{boolean flag to include n-sizes in output.  DEFAULT FALSE}

\item{ci}{boolean flag to include lower- and upper- bounds of confidence intervals.  DEFAULT FALSE}

\item{to_factor}{boolean flag to convert v1/v2 to factors, if needed.  DEFAULT TRUE}
}
\value{
dataframe of tab results
}
\description{
Run weighted or unweighted crosstabs on one or two variables, using unquoted inputs.
}
\seealso{
\code{\link{stab}} which this function wraps
}
