% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_functions.R
\name{utab}
\alias{utab}
\title{Run unweighted crosstabs on one or two variables, using quoted inputs.}
\usage{
utab(df, v1, v2 = "NULL", nsize = FALSE, ci = FALSE)
}
\arguments{
\item{df}{input dataframe}

\item{v1}{string name of 1st variable/column to tabulate}

\item{v2}{string name of optional 2nd variable/column to tablulate.  DEFAULT "NULL"}

\item{nsize}{boolean flag to include n-sizes in output.  DEFAULT FALSE}

\item{ci}{boolean flag to include lower- and upper- bounds of confidence intervals.  DEFAULT FALSE}
}
\value{
dataframe of tab results
}
\description{
Run unweighted crosstabs on one or two variables, using quoted inputs.
}
