% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_functions.R
\name{valid_tab_inputs}
\alias{valid_tab_inputs}
\title{Valid Tab Inputs}
\usage{
valid_tab_inputs(df, v1, v2 = "NULL")
}
\arguments{
\item{df}{dataframe to check}

\item{v1}{variable name to check}

\item{v2}{variable name to check.  DEFAULT "NULL"}
}
\value{
error/warning message, or empty string
}
\description{
Check input dataframe and column(s) to tab functions.
}
